Remove spot market + optimisations (#2146)
* Remove spot markets from contracts
- synth id -> collateral address
- skew scale set internally on each collateral
- Use immutables to grab sUSD address

* Update tests and bootstrap

* Remove spot market from pgk json

* Update storage

* Remove spot market test

* Make sure core staked amount/ price is correct.

* Update comment


* Gas optimisations (#2147)

* Only grab sUSD from immutables

* Use immutable pattern to set oracle manager address
- This avoids a lot of storage reads

* Use immutable pattern to set synthetix manager address
- This avoids a lot of storage reads

* Update storage

* Fix issues from merged PR not using the new structure

* Fix stack to deep and remove unused parameters

* refactor: rm createValidatedTrade

"Stack too deep" may have been due to complex inline computation. Storing
nextMarginUsd and even marginUsdForNextMarginUsd separately made it such that
we could more predictably control stack usage.

Hence createValidatedTrade is not necessary.

* Only pass in Core address and lookup sUSD and oracle

* Update comment


* typo

* Inline approvals

* Remove unnecessary vars

* Pass in globalMarginConfig instead of loading twice

* Address registry (#2150)

* Use AddressRegistry

* Create getHealthFactor and resolve some more stack too deep

* Missed commiting these changes for some reason

* Use synthetix variable


---------


* Stop relying on runtime.i

* Remove Runtime_getCollateralUsd

* Remove structs from `Runtime_validateTrade`

* Remove `Runtime_commitOrder` and remove ValidatedTrade from `Runtime_settleOrder`

* Update storage dump

* update cannon pre release

* v3.3.18

* Testnet deployment fixes (#2153)

* Make marketId part of ConfigureByMarketParameters

* Add a call for `setRewardDistributorImplementation` in the production cannonfile.
- We already have this in the test cannonfile

* Bfp market opt simplified (#2152)

* refactor: rm unused config and shuffle struct attrs

* refactor: pull AVG_SECONDS_PER_YEAR to const

* feat: uint128 keeperFeeBufferUsd and uint64 commitmentTime

* perf: add padding for Settlement{...}.GlobalData.maxHooksPerOrder

* refactor: constructor, comments, up storage.dump __unused{n}

* refactor: minimise block.timestamp and shuffle PerpMarket.Data

* perf: uint256->int128 for skew, funding, velocity, util

* refactor: more uint64 for commitmentTime

* refactor+perf+docs: constants and comments

* refactor: rm unused imports, libs, and explicit conversions

* refactor: rm unused custom errors

* refactor: minor rename and formatting in tests

* perf: represent {min,max}OrderAge as uint64

* Use alpha version

---------

+++ markets/bfp-market/contracts/interfaces/IOrderModule.sol
@@ -12,11 +12,11 @@ interface IOrderModule is IBasePerpMarket {
         /// Size to modify when settled
         int128 sizeDelta;
         /// block.timestamp of when the order was committed
-        uint256 commitmentTime;
+        uint64 commitmentTime;
         /// The max acceptable price tolerance for settlement
         uint256 limitPrice;
         /// A tip in USD to pay for settlement keepers
-        uint256 keeperFeeBufferUsd;
+        uint128 keeperFeeBufferUsd;
         /// A list of whitelisted hook addresses to invoke after settlement
         address[] hooks;
         /// True if order expired and must be canceled, false otherwise
@@ -37,7 +37,7 @@ interface IOrderModule is IBasePerpMarket {
     event OrderCommitted(
         uint128 indexed accountId,
         uint128 indexed marketId,
-        uint256 commitmentTime,
+        uint64 commitmentTime,
         int128 sizeDelta,
         uint256 estimatedOrderFee,
         uint256 estimatedKeeperFee
@@ -57,12 +57,12 @@ interface IOrderModule is IBasePerpMarket {
     event OrderSettled(
         uint128 indexed accountId,
         uint128 indexed marketId,
-        uint256 settlementTime,
+        uint64 settlementTime,
         int128 sizeDelta,
         uint256 orderFee,
         uint256 keeperFee,
-        int256 accruedFunding,
-        uint256 accruedUtilization,
+        int128 accruedFunding,
+        uint128 accruedUtilization,
         int256 pnl,
         uint256 fillPrice,
         uint128 accountDebt
@@ -92,7 +92,7 @@ interface IOrderModule is IBasePerpMarket {
         uint128 marketId,
         int128 sizeDelta,
         uint256 limitPrice,
-        uint256 keeperFeeBufferUsd,
+        uint128 keeperFeeBufferUsd,
         address[] memory hooks
     ) external;
 
@@ -150,7 +150,7 @@ interface IOrderModule is IBasePerpMarket {
     function getOrderFees(
         uint128 marketId,
         int128 sizeDelta,
-        uint256 keeperFeeBufferUsd
+        uint128 keeperFeeBufferUsd
     ) external view returns (uint256 orderFee, uint256 keeperFee);
 
     /// @notice Returns an oracle price adjusted by a premium/discount based on how the sizeDelta effects skew.
